Consolidate Authorization extensions
