+++ sources/Google.Solutions.IapDesktop.Extensions.Shell/Views/Credentials/CreateCredentialsWorkflow.cs
@@ -24,6 +24,7 @@ using Google.Solutions.IapDesktop.Application.ObjectModel;
 using Google.Solutions.IapDesktop.Application.Services.Authorization;
 using Google.Solutions.IapDesktop.Application.Services.Integration;
 using Google.Solutions.IapDesktop.Application.Services.Windows;
+using Google.Solutions.IapDesktop.Extensions.Shell.Data;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.ConnectionSettings;
 using System;
 using System.Threading.Tasks;
@@ -60,7 +61,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Shell.Views.Credentials
             bool silent)
         {
             var username = string.IsNullOrEmpty(settings.RdpUsername.StringValue)
-                ? this.serviceProvider.GetService<IAuthorization>().SuggestWindowsUsername()
+                ? SamAccountName.SuggestFromGoogleEmailAddress(this.serviceProvider.GetService<IAuthorization>())
                 : settings.RdpUsername.StringValue;
 
             if (!silent)
